Fix cyclic dependency
